server/requirements.txt
Flask==1.1.1
mongoengine==0.18.2
python-dotenv==0.10.3